Fix compilation on Debian
json-c 0.13 isn't required.
The size_t -> int changes are more correct as that's what the json functions return.

+++ sway/ipc-server.c
@@ -524,7 +524,7 @@ void ipc_client_handle_command(struct ipc_client *client) {
 		}
 
 		// parse requested event types
-		for (size_t i = 0; i < json_object_array_length(request); i++) {
+		for (int i = 0; i < json_object_array_length(request); i++) {
 			const char *event_type = json_object_get_string(json_object_array_get_idx(request, i));
 			if (strcmp(event_type, "workspace") == 0) {
 				client->subscribed_events |= event_mask(IPC_EVENT_WORKSPACE);
